requirements.txt
tk
mysql-connector-python